refactor: task status, remove uuid package
