+++ src/tasks/dto/get-tasks.dto.ts
@@ -1,5 +1,5 @@
 import { IsEnum, IsNotEmpty, IsOptional, IsString } from "class-validator";
-import { TaskStatus } from "../task.model";
+import { TaskStatus } from "../task-status.enum";
 
 export class GetTasksFilterDto {
     @IsOptional()
