--- src/tasks/task.model.ts
@@ -1,12 +1,0 @@
-export interface Task {
-    id: string;
-    title: string;
-    description: string;
-    status: TaskStatus;
-}
-
-export enum TaskStatus {
-    OPEN = 'OPEN',
-    IN_PROGRESS = 'IN_PROGRESS',
-    DONE = 'DONE'
-}
